+++ lib/ortho/sgraph.h
@@ -57,9 +57,9 @@ typedef struct {
 
 extern void reset(sgraph*);
 extern void gsave(sgraph*);
-extern sgraph* createSGraph(int);
+extern sgraph *createSGraph(size_t);
 extern void freeSGraph (sgraph*);
-extern void initSEdges (sgraph* g, int maxdeg);
+extern void initSEdges(sgraph *g, size_t maxdeg);
 extern int shortPath (sgraph* g, snode* from, snode* to);
 extern snode* createSNode (sgraph*);
 extern sedge* createSEdge (sgraph* g, snode* v0, snode* v1, double wt);
